+++ Installer/OS/ubuntu24_04.py
@@ -105,6 +105,7 @@ sudo python3 -m pip install msgpack --break-system-packages
 sudo python3 -m pip install eventlet --break-system-packages
 sudo python3 -m pip install psycopg2-binary --break-system-packages
 sudo python3 -m pip install python-dotenv --break-system-packages
+sudo apt-get install -y iw
 
 sudo apt-get install -y gnome-terminal dbus-x11 gsettings-desktop-schemas  # Apptainer